golangci-lint/uncovert: Address recommendation

+++ pkg/utils/directory.go
@@ -12,5 +12,5 @@ type Directory struct {
 }
 
 func timespecToTime(ts syscall.Timespec) time.Time {
-	return time.Unix(int64(ts.Sec), int64(ts.Nsec))
+	return time.Unix(ts.Sec, ts.Nsec)
 }
